refactor: #183 Added gradient’s exports & sorted gradients
