+++ src/lib/atoms/mesh-gradients/meshgrad-blue-static.svelte
