+++ src/lib/atoms/mesh-gradients/meshgrad-green-static.svelte
