+++ src/lib/atoms/mesh-gradients/meshgrad-pink-static.svelte
