+++ src/lib/atoms/mesh-gradients/meshgrad-red-static.svelte
