+++ src/lib/index.js
@@ -1,18 +1,32 @@
 // place files you want to import through the `$lib` alias in this folder.
-export { default as MeshgradBlue } from './meshgrad-blue.svelte'
-export { default as MeshgradRed } from './meshgrad-red.svelte'
-export { default as MeshgradGreen } from './meshgrad-green.svelte'
-export { default as MeshgradPink } from './meshgrad-pink.svelte'
-export { default as BlurgradBlue } from './blurgrad-blue.svelte'
-export { default as BlurgradRed } from './blurgrad-red.svelte'
-export { default as BlurgradGreen } from './blurgrad-green.svelte'
-export { default as BlurgradPink } from './blurgrad-pink.svelte'
+// Atoms
+export { default as MeshgradBlue } from './atoms/mesh-gradients/meshgrad-blue.svelte'
+export { default as MeshgradBlueStatic } from './atoms/mesh-gradients/meshgrad-blue-static.svelte'
+export { default as MeshgradRed } from './atoms/mesh-gradients/meshgrad-red.svelte'
+export { default as MeshgradRedStatic } from './atoms/mesh-gradients/meshgrad-red-static.svelte'
+export { default as MeshgradGreen } from './atoms/mesh-gradients/meshgrad-green.svelte'
+export { default as MeshgradGreenStatic } from './atoms/mesh-gradients/meshgrad-green-static.svelte'
+export { default as MeshgradPink } from './atoms/mesh-gradients/meshgrad-pink.svelte'
+export { default as MeshgradPinkStatic } from './atoms/mesh-gradients/meshgrad-pink-static.svelte'
+
+export { default as BlurgradBlue } from './atoms/blur-gradients/blurgrad-blue.svelte'
+export { default as BlurgradRed } from './atoms/blur-gradients/blurgrad-red.svelte'
+export { default as BlurgradGreen } from './atoms/blur-gradients/blurgrad-green.svelte'
+export { default as BlurgradPink } from './atoms/blur-gradients/blurgrad-pink.svelte'
+
 export { default as ArrowL } from './arrow-l.svelte'
 export { default as ArrowR } from './arrow-r.svelte'
+export { default as NavIcons } from './atoms/icons-nav.svelte'
+
+// Components
+export { default as Card } from './components/Card.svelte'
+export { default as Header } from './components/Header.svelte'
+export { default as Nav } from './components/Nav.svelte'
 
 // intro-form components
 export { default as QuestionForm } from './organism/question-form.svelte'
 export { default as QuestionOption } from './atoms/question-option.svelte'
 export { default as QuestionSet } from './molecules/question-set.svelte'
 
-export { default as NavIcons } from './atoms/icons-nav.svelte'
+export { default as NavButtons } from './NavButtons.svelte'
+export { default as HeaderIntro } from './HeaderIntro.svelte'
